Calculate codes more efficiently
Raise the InvalidArgument exception based on the values of codes. This is
just slightly better organization, because the exception is raised only
if the blacklist actually leads to an impossible to satisfy strategy.

This also makes codes have a list type. In the previous version codes
could be a tuple or a list, which is harder to type check.

Calculate frozenset(blacklist) only once, and use set operations for
clarity.

Signed-off-by: mulhern <amulhern@redhat.com>